Fix BigInteger ForeignKey. Added alembic cfg. Edit handlers. Added keyword filter.

+++ core/database/migrations/versions/1ff9c5f5fc53_init.py
@@ -28,7 +28,7 @@ def upgrade() -> None:
                     sa.Column('chat_id', sa.BigInteger(), nullable=False),
                     sa.Column('chat_title', sa.String(), nullable=False),
                     sa.Column('keywords', sa.Text(), nullable=True),
-                    sa.Column('owner_id', sa.Integer(), nullable=False),
+                    sa.Column('owner_id', sa.BigInteger(), nullable=False),
                     sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
                     sa.PrimaryKeyConstraint('id'),
                     sa.UniqueConstraint('chat_id')
